MNSIT  HANDWRITING RECOGNITION MODEL
Welcome to the MNIST Handwriting Recognition Project repository! This project focuses on building a machine learning model to recognize handwritten digits using the famous MNIST dataset. The MNIST dataset consists of a vast collection of 28x28 pixel grayscale images of handwritten digits (0 through 9), making it an ideal benchmark for beginners and experts alike in the field of image classification.